the first attempts of MVVM using ViewModel of arch.lifecycle library

+++ app/src/main/java/com/example/ilya4/retrofitsimpleweather/pojo/Wind.java
@@ -5,6 +5,7 @@ import java.io.Serializable;
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
+
 public class Wind implements Serializable
 {
 
@@ -13,7 +14,7 @@ public class Wind implements Serializable
     private Double speed;
     @SerializedName("deg")
     @Expose
-    private Integer deg;
+    private Double deg;
     private final static long serialVersionUID = -1790052897136013146L;
 
     public Double getSpeed() {
@@ -24,11 +25,11 @@ public class Wind implements Serializable
         this.speed = speed;
     }
 
-    public Integer getDeg() {
+    public Double getDeg() {
         return deg;
     }
 
-    public void setDeg(Integer deg) {
+    public void setDeg(Double deg) {
         this.deg = deg;
     }
